Route Table Prototype
Built a prototype of the route table, which is a data structure that
maps http routes, ie `/index.html`, `/`, `/images/image.png` to
resolvers, which are essentially functions that when called, produce an
http response.

Resolvers are represented as an Enum with a resolve() function. Each
member of the enum represents a different set of logic, and can
optionally store data. At the moment, these are

- Text: for text files
- Html: for html files
- Directory: for serving whole directories
- Executable: for serving the results of running an executable file
- Markdown: for serving markdown files that have been rendered to html
- Embedded: for functions built into or added on to the serv executable

This is a very rough draft of what I want resolvers to look like, I
expect it to change significantly as the project moves forward.

Routing is done using tide's server with `server.at("")` and
`server.at("*route")` both pointing to a handler function that does its
own routing based on the route table and the route parameter.