Slight Refactor
Moved the Resolver enum to its own file.
Moved the server logic back into main.rs

Created a working File Resolver, although currently it only works with
absolute paths. In the future this should be more flexible. I also want
to add a field to `Resolver::File` to track its file type, mainly so
that it can set the appropriate http header in the response.