Added the Exec resolver
A route with the Exec resolver can point to any executable file on the
host system. When called, it will run the executable and collect its
standard output to be returned as text. The idea is that this could
point to scripts in any language, and could be used to build simple APIs
extremely quickly and with very little boilerplate.

The included `public/shell_example` file demonstrates some of what can be done.

I'm using the
[async-process](https://docs.rs/async-process/1.1.0/async_process/)
crate to handle all of the external process calls. Hopefully, by doing
this asynchronously, the risk of a hanging process killing the server
can be reduced.

TODO: I need to sort out a good way of accepting arguments and passing
them to the executables.