First Implementation of a Routes File
My hope for the routes file is that it will more or less completely
configure an instance of serv, in as concise a way possible.

Each route represents a link between a potential HTTP request and a
resource on the host machine. One route is defined on every line of the
route file. A line consists of the http request, followed by the path of
the associated resource on the host machine (relative to the directory
of the routes file) followed by a symbol representing the intended
resolver method, separated by any amount of whitespace.

Currently, the `f` symbol is used to indicate the `file` resolver,
and the `x` symbol indicates the `exec` resolver. Although I suspect
this might change in the future.

The example below contains 3 routes. An html file, which uses the file
resolver, and a shell and nodejs script, which use the execute resolver.

```
/        index.html     f
/shell   shell_example  x   # comments can be inline
/js      js             x
```