Executable Files Can Now Accept Arguments
After the exec keyword in the routes file, the user can specify
arguments, such as elements of the url query and indices of the
wildcards matched against in the path. For example, in the route

`/one/two/*    scripts/render.sh    exec(wild:0, query:type)`

The `render.sh` script will receive two arguments upon execution. $1
will be the first wildcard, and $2 will be the part of the url query
with the key `type`.

For example if the request `http://localhost:4000/one/two/three.md?type=markdown` is
matched, then the `render.sh` will receive arguments `three.md` and
`markdown`