Refactor of resolvers.rs file to options.rs
In the routes file, the text following the request and resource paths on
each line is now called the "options list", which the user can fill with
zero or more options.

Options can be either an "access type" or a "post processor". Access
Types work like functions which take the http request as an input and
return a Response object,
`(Request) -> (Response)`
While post processors both take and return the Response type,
`(Response) -> (Response)`

The options list must consist of exactly one Access Type followed by
zero or more Post Processors, such that when they are composed together,
they have the type `(Request) -> (Response)`. If no access type is
specified by the user, it is assumed to be Read and inserted
automatically

---

I am a little bit torn over the usefulness of these post processors. On
the one hand, the ability to execute arbitrary shell scripts more or
less covers all of the uses for these I can think of. For example I
thought about the idea of adding options like image compression or
encryption, but I think in most cases it would simply be easier for the
route to invoke a shell script that calls the necessary programs for
you.

FossilOrigin-Name: cb8df28b0287b2b3ebeefcad66eb260e8211c3aca0c186980a3eb7607bbe747c